Small fix in ModelTest to test if the test notebook directory exists

+++ src/test/kotlin/ModelTest.kt
@@ -27,8 +27,10 @@ internal class ModelTest {
             "</head><body contenteditable=\"true\">Hello <a href=\"https://github.com/TestFX/TestFX\">world</a></body></html>"
 
     private fun clearTestNotebookDirectory() {
-        for (file in testNotebookDir.listFiles()) {
-            file.delete()
+        if (testNotebookDir.exists()) {
+            for (file in testNotebookDir.listFiles()) {
+                file.delete()
+            }
         }
     }
